+++ spec/acts_as_hashed_spec.rb
@@ -10,6 +10,8 @@ FileUtils.rm_f DB_FILE
 ActiveRecord::Base.establish_connection :adapter => 'sqlite3', :database => DB_FILE
 ActiveRecord::Base.connection.execute 'CREATE TABLE some_models (id INTEGER NOT NULL PRIMARY KEY, hashed_code string)'
 ActiveRecord::Base.connection.execute 'CREATE TABLE overwrited_models (id INTEGER NOT NULL PRIMARY KEY, hashed_code string)'
+ActiveRecord::Base.connection.execute 'CREATE TABLE custom_callback_models (id INTEGER NOT NULL PRIMARY KEY, hashed_code string)'
+ActiveRecord::Base.connection.execute 'CREATE TABLE invalid_exception_models (id INTEGER NOT NULL PRIMARY KEY, hashed_code string)'
 
 class SomeModel < ActiveRecord::Base
   acts_as_hashed
@@ -25,20 +27,30 @@ class OverwritedModel < ActiveRecord::Base
   end
 end
 
+class CustomCallbackModel < ActiveRecord::Base
+  acts_as_hashed :before_validation
+end
+
 describe ActsAsHashed do
   before(:each) do
     clean_database!
   end
+
   describe :save do
+
     context "new record" do
       let(:model) { SomeModel.new }
-      it "should save hashed code" do
-        SomeModel.stub(:friendly_token).and_return('new-hashed-code-here')
-        expect {
-          model.save!
-        }.to change(model, :hashed_code).from(nil).to('new-hashed-code-here')
+
+      context 'when hashed_code is nil' do
+        it "should save hashed code" do
+          SomeModel.stub(:friendly_token).and_return('new-hashed-code-here')
+          expect {
+            model.save!
+          }.to change(model, :hashed_code).from(nil).to('new-hashed-code-here')
+        end
       end
     end
+
     describe "persisted record" do
       let(:model) do
         model = SomeModel.new
@@ -52,6 +64,7 @@ describe ActsAsHashed do
         }.to_not change(model, :hashed_code)
       end
     end
+
     context "overwrite friendly_token" do
       let!(:model) do
         model = OverwritedModel.new
@@ -62,7 +75,48 @@ describe ActsAsHashed do
         model.hashed_code.length.should == 10
       end
     end
+
+    context "a custom callback ('before_validation') is used" do
+      context "when validations are disabled" do
+        let!(:model) do
+          model = CustomCallbackModel.new
+          model.save(validate: false)
+          model
+        end
+
+        subject { model.hashed_code }
+
+        it "returns nil" do
+          expect(subject).to be_nil
+        end
+      end
+
+      context "when validations are enabled" do
+        let!(:model) do
+          model = CustomCallbackModel.new
+          model.save
+          model
+        end
+
+        subject { model.hashed_code }
+
+        it "returns something" do
+          expect(subject).not_to be_nil
+        end
+      end
+
+      context "when an invalid callback is used" do
+        it "raises an error" do
+          expect do
+            class InvalidExceptionModel < ActiveRecord::Base
+              acts_as_hashed :invalid_callback
+            end
+          end.to raise_error(ArgumentError)
+        end
+      end
+    end
   end
+
   describe :update_missing_hashed_code do
     let!(:model) do
       model = SomeModel.new